+++ RELEASES.md
@@ -19,6 +19,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * prov-template-compiler: generation of insert_plead_transforming_composite_composite_and_linker sql method
 * prov-template-compiler: search records function
 * service-template: search records functionality
+* service-template: browser tab